Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -3,7 +3,7 @@
 int main() {
     float a, b;
     scanf("%f %f", &a, &b);
-    float product;
+    int product;
     product = (int)a * (int)b;
     printf("Product: %d", product);
